requirements.txt
fastapi
pydantic
requests
uvicorn
openai
python-dotenv 